financial-ai-advisor/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Providers
openai>=1.12.0
anthropic==0.7.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3

# Data & Storage
sqlalchemy==2.0.23
redis==5.0.1

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Streamlit & Visualization
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
numpy<2.0.0  # Compatibility with plotly/xarray

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1

